move configure() to separated module
